Was: Merge branch 'master' of /srv/bindings/5663f3f3d4a4409c940426463a066b26/repository into joe

+++ wp-content/plugins/squirrly-seo/controllers/SQ_Frontend.php
@@ -175,7 +175,7 @@ class SQ_Frontend extends SQ_FrontController {
                     }
                     break;
                 case 'touchicon':
-                    $size = get_query_var('sq_size');
+                    $size = (int) get_query_var('sq_size');
                     if (SQ_Tools::$options['favicon'] <> '') {
                         //show the favico file
                         SQ_Tools::setHeader('png');
