+++ wp-content/plugins/squirrly-seo/models/SQ_Frontend.php
@@ -1,7 +1,6 @@
 <?php
 
-class Model_SQ_Frontend
-{
+class Model_SQ_Frontend {
 
     public $buffer;
 
@@ -43,8 +42,7 @@ class Model_SQ_Frontend
     /** @var array Meta custom content */
     private $meta = array();
 
-    public function __construct()
-    {
+    public function __construct() {
         SQ_ObjController::getController('SQ_Tools', false);
         $this->post_type = SQ_Tools::$options['sq_post_types'];
     }
@@ -55,8 +53,7 @@ class Model_SQ_Frontend
      * Write the signature
      * @return string
      */
-    public function setStart()
-    {
+    public function setStart() {
         return "\n\n<!-- Squirrly SEO Plugin " . SQ_VERSION . ", visit: http://plugin.squirrly.co/ -->\n";
     }
 
@@ -64,8 +61,7 @@ class Model_SQ_Frontend
      * Set the line where Squirrly will start the code
      * @return string
      */
-    public function setStartTag()
-    {
+    public function setStartTag() {
 
         if ($this->is_squirrly()) {
             global $post;
@@ -87,8 +83,7 @@ class Model_SQ_Frontend
      * End the signature
      * @return string
      */
-    public function setEnd()
-    {
+    public function setEnd() {
         return "<!-- /Squirrly SEO Plugin -->\n\n";
     }
 
@@ -98,8 +93,7 @@ class Model_SQ_Frontend
      * Start the buffer record
      * @return type
      */
-    public function startBuffer()
-    {
+    public function startBuffer() {
         ob_start(array($this, 'getBuffer'));
     }
 
@@ -109,8 +103,7 @@ class Model_SQ_Frontend
      * @param buffer $buffer
      * @return buffer
      */
-    public function getBuffer($buffer)
-    {
+    public function getBuffer($buffer) {
         if (isset($this->buffer)) {
             return $this->buffer;
         }
@@ -125,8 +118,7 @@ class Model_SQ_Frontend
      * @return string
      *
      */
-    public function flushHeader()
-    {
+    public function flushHeader() {
         $buffers = array();
 
 
@@ -148,8 +140,7 @@ class Model_SQ_Frontend
         }
     }
 
-    public function is_squirrly()
-    {
+    public function is_squirrly() {
         if (SQ_Tools::getValue('sq_use') == 'off') {
             return false;
         }
@@ -170,8 +161,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    public function setMetaInBuffer($buffer)
-    {
+    public function setMetaInBuffer($buffer) {
         global $post;
 
         //if the title is already shown
@@ -258,8 +248,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    public function getHeader()
-    {
+    public function getHeader() {
         $ret = '';
         $ret .= $this->setStart();
 
@@ -314,14 +303,12 @@ class Model_SQ_Frontend
         return $ret;
     }
 
-    public function setPost($newpost)
-    {
+    public function setPost($newpost) {
         global $post;
         $this->post = $post = $newpost;
     }
 
-    private function getTwitterCard()
-    {
+    private function getTwitterCard() {
         $meta = "\n";
 
         //Title and Description is required
@@ -355,8 +342,7 @@ class Model_SQ_Frontend
      * @param string $account
      * @return string | false
      */
-    public function getTwitterAccount($account)
-    {
+    public function getTwitterAccount($account) {
         if ($account <> '') {
             if (strpos($account, 'twitter.com') !== false) {
                 preg_match('/twitter.com\/([@1-9a-z_-]+)/i', $account, $result);
@@ -379,8 +365,7 @@ class Model_SQ_Frontend
      * Get the Open Graph Protocol
      * @return string
      */
-    private function getOpenGraph()
-    {
+    private function getOpenGraph() {
         $meta = "\n";
         $image = '';
 
@@ -456,10 +441,10 @@ class Model_SQ_Frontend
 
                         $meta .= sprintf('<meta property="product:sale_price:amount" content="%s" />', $product->sale_price) . "\n";
                         $meta .= sprintf('<meta property="product:sale_price:currency" content="%s" />', get_woocommerce_currency()) . "\n";
-                        if($sales_price_from > 0) {
+                        if ($sales_price_from > 0) {
                             $meta .= sprintf('<meta property="product:sale_price:start" content="%s" />', date("Y-m-d H:i:s", $sales_price_from)) . "\n";
                         }
-                        if($sales_price_to) {
+                        if ($sales_price_to) {
                             $meta .= sprintf('<meta property="product:sale_price:end" content="%s" />', date("Y-m-d H:i:s", $sales_price_to)) . "\n";
                         }
 
@@ -504,8 +489,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function setCanonical()
-    {
+    private function setCanonical() {
 
         if ($this->url) {
             remove_action('wp_head', 'rel_canonical');
@@ -516,8 +500,7 @@ class Model_SQ_Frontend
         return '';
     }
 
-    public function setRelPrevNext()
-    {
+    public function setRelPrevNext() {
         global $paged;
         $meta = "";
         if (!$this->isHomePage()) {
@@ -532,8 +515,7 @@ class Model_SQ_Frontend
         return (($meta <> '') ? apply_filters('sq_prevnext_meta', $meta) . "\n" : '');
     }
 
-    public function getTitle()
-    {
+    public function getTitle() {
         $this->getCustomTitle();
         return $this->title;
     }
@@ -543,8 +525,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    public function getCustomTitle()
-    {
+    public function getCustomTitle() {
         $sep = ' | ';
 
         if (isset($this->title)) {
@@ -561,7 +542,7 @@ class Model_SQ_Frontend
                     $this->title = $this->grabTitleFromPost();
                 }
                 if (is_paged()) {
-                    $this->title .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . get_query_var('paged');
+                    $this->title .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . (int) get_query_var('paged');
                 }
             } elseif (is_author()) { //for author
                 if ($this->title == '') {
@@ -571,12 +552,12 @@ class Model_SQ_Frontend
                     $this->title = __('About') . " " . ucfirst($this->getAuthor('display_name'));
                 }
                 if (is_paged()) {
-                    $this->title .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . get_query_var('paged');
+                    $this->title .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . (int)get_query_var('paged');
                 }
             } elseif (is_tag()) { //for tags
                 if (is_paged()) {
                     $tag = get_query_var('tag');
-                    $this->title = ucfirst(str_replace('-', ' ', $tag)) . $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . get_query_var('paged');
+                    $this->title = ucfirst(str_replace('-', ' ', $tag)) . $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . (int)get_query_var('paged');
                 }
             } elseif (is_archive()) { //for archive and products
                 if (isset($this->post) && isset($this->post->ID)) {
@@ -598,7 +579,7 @@ class Model_SQ_Frontend
                 }
 
                 if (is_paged()) {
-                    $this->title .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . get_query_var('paged');
+                    $this->title .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . (int)get_query_var('paged');
                 }
             } elseif (is_single() || is_page() || is_singular() || in_array(get_post_type(), $this->post_type)) {
                 if (isset($this->post) && isset($this->post->ID)) {
@@ -640,8 +621,7 @@ class Model_SQ_Frontend
         return apply_filters('sq_title', $this->title);
     }
 
-    public function clearTitle($title)
-    {
+    public function clearTitle($title) {
         if ($title <> '') {
             $title = SQ_Tools::i18n(trim(esc_html(ent2ncr(strip_tags($title)))));
         }
@@ -654,8 +634,7 @@ class Model_SQ_Frontend
      * @param integer $id Post ID
      * @return type
      */
-    public function getImagesFromContent($id = null, $all = false)
-    {
+    public function getImagesFromContent($id = null, $all = false) {
         $images = array();
         if (isset($id)) {
             $post = get_post($id);
@@ -717,8 +696,7 @@ class Model_SQ_Frontend
      * @param integer $id Post ID
      * @return type
      */
-    public function getVideosFromContent($id = null)
-    {
+    public function getVideosFromContent($id = null) {
         $videos = array();
 
         if (isset($id)) {
@@ -770,8 +748,7 @@ class Model_SQ_Frontend
         return $videos;
     }
 
-    public function getDescription()
-    {
+    public function getDescription() {
         $this->getCustomDescription();
         return $this->description;
     }
@@ -781,8 +758,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    public function getCustomDescription()
-    {
+    public function getCustomDescription() {
 
         $sep = ' | ';
         $description = '';
@@ -801,7 +777,7 @@ class Model_SQ_Frontend
                 }
 
                 if (is_paged()) {
-                    $description .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . get_query_var('paged');
+                    $description .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . (int)get_query_var('paged');
                 }
 
                 if ($this->isHomePage() && $description <> '') {
@@ -815,7 +791,7 @@ class Model_SQ_Frontend
                     $description = $this->grabDescriptionFromPost() . $sep . $this->getAuthor('display_name');
                 }
                 if (is_paged()) {
-                    $description .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . get_query_var('paged');
+                    $description .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . (int)get_query_var('paged');
                 }
             } elseif (is_tag()) { //for tags
                 $description = tag_description();
@@ -824,7 +800,7 @@ class Model_SQ_Frontend
                     $description = ucfirst($tag) . $sep . $this->grabDescriptionFromPost();
                 }
                 if (is_paged()) {
-                    $description .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . get_query_var('paged');
+                    $description .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . (int)get_query_var('paged');
                 }
             } elseif (is_archive()) { //for archive and products
                 if (isset($this->post) && isset($this->post->ID)) {
@@ -850,7 +826,7 @@ class Model_SQ_Frontend
                 }
 
                 if (is_paged()) {
-                    $description .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . get_query_var('paged');
+                    $description .= $sep . __('Page', _SQ_PLUGIN_NAME_) . " " . (int)get_query_var('paged');
                 }
             } elseif (is_single() || is_page() || is_singular() || $this->checkPostsPage() || in_array(get_post_type(), $this->post_type)) {
                 if (isset($this->post) && isset($this->post->ID)) {
@@ -893,8 +869,7 @@ class Model_SQ_Frontend
         return '';
     }
 
-    public function clearDescription($description)
-    {
+    public function clearDescription($description) {
         if ($description <> '') {
             $search = array("'<script[^>]*?>.*?<\/script>'si", // strip out javascript
                 "/<form.*?<\/form>/si",
@@ -915,8 +890,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getCustomKeyword()
-    {
+    private function getCustomKeyword() {
         $keywords = '';
 
         if ($this->checkPostsPage() && SQ_Tools::$options['sq_auto_description'] == 1) {
@@ -950,8 +924,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getCopyright()
-    {
+    private function getCopyright() {
         $meta = '';
 
         $name = $this->getAuthor('display_name');
@@ -971,8 +944,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getGooglePlusMeta()
-    {
+    private function getGooglePlusMeta() {
         $meta = '';
         $author = SQ_Tools::$options['sq_google_plus'];
 
@@ -992,8 +964,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getFavicon()
-    {
+    private function getFavicon() {
         $meta = '';
         $rnd = '';
 
@@ -1036,8 +1007,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getLanguage()
-    {
+    private function getLanguage() {
         $meta = '';
         $language = get_bloginfo('language');
 
@@ -1059,8 +1029,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getDublinCore()
-    {
+    private function getDublinCore() {
         $date = null;
         $meta = '';
 
@@ -1094,8 +1063,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getXMLSitemap()
-    {
+    private function getXMLSitemap() {
         $meta = '';
 
         $xml_url = SQ_ObjController::getController('SQ_Sitemaps')->getXmlUrl('sitemap');
@@ -1112,8 +1080,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getGoogleWT()
-    {
+    private function getGoogleWT() {
         $sq_google_wt = SQ_Tools::$options['sq_google_wt'];
 
         if ($this->isHomePage() && $sq_google_wt <> '') {
@@ -1128,8 +1095,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getGoogleAnalytics()
-    {
+    private function getGoogleAnalytics() {
         $sq_google_analytics = SQ_Tools::$options['sq_google_analytics'];
 
         if ($sq_google_analytics <> '') {
@@ -1156,8 +1122,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getFacebookIns()
-    {
+    private function getFacebookIns() {
         $sq_facebook_insights = SQ_Tools::$options ['sq_facebook_insights'];
 
         if ($this->isHomePage() && $sq_facebook_insights <> '') {
@@ -1172,8 +1137,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getPinterest()
-    {
+    private function getPinterest() {
         $sq_pinterest = SQ_Tools::$options['sq_pinterest'];
 
         if ($this->isHomePage() && $sq_pinterest <> '') {
@@ -1188,8 +1152,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getAlexaT()
-    {
+    private function getAlexaT() {
         $sq_alexa = SQ_Tools::$options['sq_alexa'];
 
         if ($this->isHomePage() && $sq_alexa <> '') {
@@ -1204,8 +1167,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    private function getBingWT()
-    {
+    private function getBingWT() {
         $sq_bing_wt = SQ_Tools::$options['sq_bing_wt'];
 
         if ($this->isHomePage() && $sq_bing_wt <> '') {
@@ -1219,8 +1181,7 @@ class Model_SQ_Frontend
      * Get the JsonLD meta for this site
      * @return string
      */
-    private function getJsonLD()
-    {
+    private function getJsonLD() {
         $meta = '';
         $sep = ",\n";
         if ($this->isHomePage()) {
@@ -1352,8 +1313,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    public function grabTitleFromPost($id = null)
-    {
+    public function grabTitleFromPost($id = null) {
         global $wp_query;
         $post = null;
         $title = '';
@@ -1393,8 +1353,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    public function grabDescriptionFromPost($id = null)
-    {
+    public function grabDescriptionFromPost($id = null) {
         global $wp_query;
         $post = null;
 
@@ -1442,8 +1401,7 @@ class Model_SQ_Frontend
      *
      * @return array
      */
-    public function grabKeywordsFromPost($id = null)
-    {
+    public function grabKeywordsFromPost($id = null) {
         global $wp_query;
 
         $this->max_keywords = ($this->max_keywords > 0 ? ($this->max_keywords - 1) : 0);
@@ -1549,8 +1507,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    public function getCanonicalUrl()
-    {
+    public function getCanonicalUrl() {
         global $wp_query;
 
         if (isset($this->url)) {
@@ -1632,8 +1589,7 @@ class Model_SQ_Frontend
         return apply_filters('sq_canonical', $link);
     }
 
-    public function getAuthor($what = 'user_nicename')
-    {
+    public function getAuthor($what = 'user_nicename') {
         if (!isset($this->author)) {
             if (is_author()) {
                 $this->author = get_userdata(get_query_var('author'));
@@ -1655,9 +1611,8 @@ class Model_SQ_Frontend
         return false;
     }
 
-    public function getPaged($link)
-    {
-        $page = get_query_var('paged');
+    public function getPaged($link) {
+        $page = (int)get_query_var('paged');
         if ($page && $page > 1) {
             $link = trailingslashit($link) . "page/" . "$page/";
         }
@@ -1669,15 +1624,13 @@ class Model_SQ_Frontend
      *
      * @return bool
      */
-    private function isHomePage()
-    {
+    private function isHomePage() {
         global $wp_query;
 
         return (is_home() || (isset($wp_query->query) && empty($wp_query->query) && !is_preview()));
     }
 
-    private function isHtmlHeader()
-    {
+    private function isHtmlHeader() {
         $headers = headers_list();
 
         foreach ($headers as $index => $value) {
@@ -1705,8 +1658,7 @@ class Model_SQ_Frontend
      *
      * @return bool
      */
-    private function checkFrontPage()
-    {
+    private function checkFrontPage() {
         return is_page() && get_option('show_on_front') == 'page' && isset($this->post->ID) && $this->post->ID == get_option('page_on_front');
     }
 
@@ -1715,13 +1667,11 @@ class Model_SQ_Frontend
      *
      * @return bool
      */
-    private function checkPostsPage()
-    {
+    private function checkPostsPage() {
         return is_home() && get_option('show_on_front') == 'page' && isset($this->post->ID) && $this->post->ID == get_option('page_for_posts');
     }
 
-    public function truncate($text, $min, $max)
-    {
+    public function truncate($text, $min, $max) {
         if (function_exists('strip_tags')) {
             $text = strip_tags($text);
         }
@@ -1738,8 +1688,7 @@ class Model_SQ_Frontend
         return trim(stripcslashes($text));
     }
 
-    public function _truncate($text)
-    {
+    public function _truncate($text) {
         if (function_exists('strip_tags'))
             $text = strip_tags($text);
         $text = str_replace(']]>', ']]&gt;', $text);
@@ -1754,8 +1703,7 @@ class Model_SQ_Frontend
      *
      * @return string
      */
-    public function getUniqueKeywords($keywords)
-    {
+    public function getUniqueKeywords($keywords) {
         $all = array();
         if (is_array($keywords)) {
             foreach ($keywords as $word) {
@@ -1785,8 +1733,7 @@ class Model_SQ_Frontend
      * @param type $post_id
      * @return boolean
      */
-    public function getAdvancedMeta($post_id, $meta = 'title')
-    {
+    public function getAdvancedMeta($post_id, $meta = 'title') {
         global $wpdb;
 
         $field = '';
@@ -1855,8 +1802,7 @@ class Model_SQ_Frontend
      * @param type $post_id
      * @return boolean
      */
-    public function getOtherPluginsMeta($post_id, $meta = 'title')
-    {
+    public function getOtherPluginsMeta($post_id, $meta = 'title') {
         global $wpdb;
 
         $field = '';
@@ -1918,8 +1864,7 @@ class Model_SQ_Frontend
      * ROBOTSTXT
      */
     // add sitemap location in robots.txt generated by WP
-    public function robots($content = '')
-    {
+    public function robots($content = '') {
         global $blog_id;
 
         /** display robots.txt */
